requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
transformers
pypdf
nump